Made fixes / refactored
